wrote unit test to check successful and unsuccesful payment

+++ PaymentGateway/PaymentGateway/CheckoutService.cs
@@ -9,7 +9,7 @@
             _paymentGateway = paymentGateway;
         }
 
-        public PaymentResult Checkout(double amount)
+        public PaymentResult Checkout(decimal amount)
         {
             return _paymentGateway.Pay(amount);
         }
